backgroundTasks/repeatedBgTasks/requirements.txt
Django==3.1.6
django-background-tasks
